moved some tests from test_coverage away to their respective files
