+++ evap/evaluation/tests/test_coverage.py
@@ -142,81 +142,3 @@ class URLTests(WebTest):
 
     def test_contributor_settings(self):
         self.get_submit_assert_302("/contributor/settings", "responsible")
-
-    def helper_semester_state_views(self, course_ids, old_state, new_state, operation):
-        page = self.app.get("/staff/semester/1", user="evap")
-        form = page.forms["form_" + old_state]
-        for course_id in course_ids:
-            self.assertIn(Course.objects.get(pk=course_id).state, old_state)
-        form['course'] = course_ids
-        response = form.submit('operation', value=operation)
-
-        form = response.forms["course-operation-form"]
-        response = form.submit()
-        self.assertIn("Successfully", str(response))
-        for course_id in course_ids:
-            self.assertEqual(Course.objects.get(pk=course_id).state, new_state)
-
-    """
-        The following tests make sure the course state transitions are triggerable via the UI.
-    """
-    def test_semester_publish(self):
-        self.helper_semester_state_views([7], "reviewed", "published", "publish")
-
-    def test_semester_reset_1(self):
-        self.helper_semester_state_views([2], "prepared", "new", "revertToNew")
-
-    def test_semester_reset_2(self):
-        self.helper_semester_state_views([4], "approved", "new", "revertToNew")
-
-    def test_semester_approve_1(self):
-        self.helper_semester_state_views([1], "new", "approved", "approve")
-
-    def test_semester_approve_2(self):
-        self.helper_semester_state_views([2], "prepared", "approved", "approve")
-
-    def test_semester_approve_3(self):
-        self.helper_semester_state_views([3], "editor_approved", "approved", "approve")
-
-    def test_semester_contributor_ready_1(self):
-        self.helper_semester_state_views([1, 10], "new", "prepared", "prepare")
-
-    def test_semester_contributor_ready_2(self):
-        self.helper_semester_state_views([3], "editor_approved", "prepared", "reenableEditorReview")
-
-    def test_semester_unpublish(self):
-        self.helper_semester_state_views([8], "published", "reviewed", "unpublish")
-
-    def test_student_vote(self):
-        """
-            Submits a student vote for coverage, verifies that an error message is
-            displayed if not all rating questions have been answered and that all
-            given answers stay selected/filled and that the student cannot vote on
-            the course a second time.
-        """
-        page = self.get_assert_200("/student/vote/5", user="lazy.student")
-        form = page.forms["student-vote-form"]
-        form["question_17_2_3"] = "some text"
-        form["question_17_2_4"] = 1
-        form["question_17_2_5"] = 6
-        form["question_18_1_1"] = "some other text"
-        form["question_18_1_2"] = 1
-        form["question_19_1_1"] = "some more text"
-        form["question_19_1_2"] = 1
-        form["question_20_1_1"] = "and the last text"
-        response = form.submit()
-
-        self.assertIn("vote for all rating questions", response)
-        form = page.forms["student-vote-form"]
-        self.assertEqual(form["question_17_2_3"].value, "some text")
-        self.assertEqual(form["question_17_2_4"].value, "1")
-        self.assertEqual(form["question_17_2_5"].value, "6")
-        self.assertEqual(form["question_18_1_1"].value, "some other text")
-        self.assertEqual(form["question_18_1_2"].value, "1")
-        self.assertEqual(form["question_19_1_1"].value, "some more text")
-        self.assertEqual(form["question_19_1_2"].value, "1")
-        self.assertEqual(form["question_20_1_1"].value, "and the last text")
-        form["question_20_1_2"] = 1  # give missing answer
-        form.submit()
-
-        self.get_assert_403("/student/vote/5", user="lazy.student")
