+++ evap/staff/tests/test_forms.py
@@ -4,7 +4,8 @@ from model_mommy import mommy
 
 from evap.evaluation.models import UserProfile, CourseType, Course, Questionnaire, \
     Contribution, Semester, Degree, EmailTemplate
-from evap.evaluation.tests.tools import get_form_data_from_instance, course_with_responsible_and_editor, to_querydict
+from evap.evaluation.tests.tools import get_form_data_from_instance, course_with_responsible_and_editor, to_querydict, \
+    WebTest
 from evap.staff.forms import UserForm, SingleResultForm, ContributionFormSet, ContributionForm, CourseForm, \
     CourseEmailForm
 from evap.contributor.forms import CourseForm as ContributorCourseForm
@@ -430,3 +431,51 @@ class CourseFormTests(TestCase):
 
         # staff: but start date must be < end date
         self.helper_date_validation(CourseForm, "02/1/1999", "02/1/1998", False)
+
+
+class TestCourseOperationForms(WebTest):
+    fixtures = ['minimal_test_data']
+
+    def helper_semester_state_views(self, course_ids, old_state, new_state, operation):
+        page = self.app.get("/staff/semester/1", user="evap")
+        form = page.forms["form_" + old_state]
+        for course_id in course_ids:
+            self.assertIn(Course.objects.get(pk=course_id).state, old_state)
+        form['course'] = course_ids
+        response = form.submit('operation', value=operation)
+
+        form = response.forms["course-operation-form"]
+        response = form.submit()
+        self.assertIn("Successfully", str(response))
+        for course_id in course_ids:
+            self.assertEqual(Course.objects.get(pk=course_id).state, new_state)
+
+    """
+        The following tests make sure the course state transitions are triggerable via the UI.
+    """
+    def test_semester_publish(self):
+        self.helper_semester_state_views([7], "reviewed", "published", "publish")
+
+    def test_semester_reset_1(self):
+        self.helper_semester_state_views([2], "prepared", "new", "revertToNew")
+
+    def test_semester_reset_2(self):
+        self.helper_semester_state_views([4], "approved", "new", "revertToNew")
+
+    def test_semester_approve_1(self):
+        self.helper_semester_state_views([1], "new", "approved", "approve")
+
+    def test_semester_approve_2(self):
+        self.helper_semester_state_views([2], "prepared", "approved", "approve")
+
+    def test_semester_approve_3(self):
+        self.helper_semester_state_views([3], "editor_approved", "approved", "approve")
+
+    def test_semester_contributor_ready_1(self):
+        self.helper_semester_state_views([1, 10], "new", "prepared", "prepare")
+
+    def test_semester_contributor_ready_2(self):
+        self.helper_semester_state_views([3], "editor_approved", "prepared", "reenableEditorReview")
+
+    def test_semester_unpublish(self):
+        self.helper_semester_state_views([8], "published", "reviewed", "unpublish")
